requirements.txt
streamlit==1.30.0
streamlit-extras
watchdog
openai
google-cloud-aiplatform
google-cloud-secret-manager
langchain
langchain-openai
langchain-google-genai
langchain-google-vertexai
google-cloud-firestore
streamlit-extras